+++ libft/printf_fd/ft_printf.h
@@ -22,7 +22,7 @@ void	ft_put_lu_nbr(unsigned int nbr, int *count, int fd);
 int		ft_count_digit(int nb);
 int		ft_putlnbr(int nbr, int fd);
 int		ft_putlstr(char *str, int fd);
-int		ft_put_i_nbr(int nbr, int fd);
+int		ft_put_i_nbr(long long int nbr, int fd);
 int		ft_printf_fd(int fd, const char *s, ...);
 void	ft_put_lx_nbr(unsigned int nbr, int *count, int fd);
 void	ft_put_bigx_nbr(unsigned int nbr, int *count, int fd);
